requirements.txt
h5py==3.10.0
matlabengine==23.2.1
matplotlib==3.8.2
numpy==1.26.3
scipy==1.12.0
torch==2.1.1
tqdm==4.66.1